# src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Transaction Risk Prediction API
  description: |
    Inference API for predicting transaction risk using XGBoost model.

    **Note:** This API accepts pre-computed feature vectors only. Raw account identifiers
    and behavioral history computation are outside scope—that responsibility belongs to an
    online feature store.
  version: "2.1.0"
  contact:
    name: Project Repository
    url: https://github.com/Samarthuday/Quantitative-Transaction-Risk-Modeling

servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.example.com
    description: Production

paths:
  /api/health:
    get:
      summary: Service health check
      description: |
        Returns the current health status of the service and model availability.
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service unavailable or model failed to load
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/model/info:
    get:
      summary: Model metadata and evaluation metrics
      description: |
        Returns information about the trained model including version,
        decision threshold, and test set performance metrics.
      operationId: getModelInfo
      tags:
        - Model
      responses:
        '200':
          description: Model information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfoResponse'
        '503':
          description: Model not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/predict:
    post:
      summary: Predict transaction risk probability
      description: |
        Predicts the risk probability for a transaction given its pre-computed feature vector.

        **Important:** Input must be a complete feature vector with all required fields.
        Missing or invalid fields will result in a 400 error.
      operationId: predict
      tags:
        - Prediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureVector'
            examples:
              transaction:
                summary: Example feature vector
                value:
                  Amount: 1000.0
                  log_amount: 6.908
                  hour_sin: 0.5
                  hour_cos: 0.866
                  dow_sin: 0.0
                  dow_cos: 1.0
                  month_sin: 0.259
                  month_cos: 0.966
                  is_weekend: 0
                  is_night: 0
                  currency_mismatch: 0
                  cross_border: 0
                  is_round_amount: 1
                  sender_txn_count_24h: 5
                  sender_amount_sum_24h: 5000.0
                  sender_amount_mean_30d: 1000.0
                  sender_amount_std_30d: 500.0
                  sender_amount_zscore: 0.0
                  receiver_txn_count_24h: 3
                  receiver_amount_sum_24h: 3000.0
                  seconds_since_sender_txn: 3600
                  sender_txn_count_lifetime: 100
                  receiver_txn_count_lifetime: 50
                  sender_out_degree: 25
                  receiver_in_degree: 20
                  pair_transaction_count: 5
                  sender_counterparty_hhi: 0.2
                  Payment_type: "Transfer"
                  Payment_currency: "USD"
                  Received_currency: "USD"
                  Sender_bank_location: "US"
                  Receiver_bank_location: "US"
      responses:
        '200':
          description: Risk prediction successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResponse'
        '400':
          description: Invalid request (missing/invalid features)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Model not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - model_loaded
        - timestamp
      properties:
        status:
          type: string
          enum:
            - healthy
            - model_unavailable
          description: Overall service health status
        model_loaded:
          type: boolean
          description: Whether the model artifact is available
        timestamp:
          type: string
          format: date-time
          description: Current UTC timestamp
        detail:
          type: string
          nullable: true
          description: Error details if model failed to load

    ModelInfoResponse:
      type: object
      required:
        - model
        - model_version
        - alert_rate
        - decision_threshold
        - test_metrics
      properties:
        model:
          type: string
          example: "XGBoost"
          description: Model type identifier
        model_version:
          type: string
          example: "2.1.0"
          description: Semantic version of the model
        alert_rate:
          type: number
          format: float
          example: 0.005
          description: Alert rate used for threshold selection (0.5%)
        decision_threshold:
          type: number
          format: float
          example: 0.156
          description: Calibrated probability threshold for alerts
        test_metrics:
          type: object
          description: Model performance metrics on held-out test set
          properties:
            pr_auc:
              type: number
              format: float
              description: Precision-Recall AUC
            roc_auc:
              type: number
              format: float
              description: ROC AUC
            brier_score:
              type: number
              format: float
              description: Brier score (lower is better)
            log_loss:
              type: number
              format: float
              description: Logarithmic loss
            alert_0.1%_precision:
              type: number
              format: float
            alert_0.1%_recall:
              type: number
              format: float
            alert_0.5%_precision:
              type: number
              format: float
            alert_0.5%_recall:
              type: number
              format: float
            alert_1.0%_precision:
              type: number
              format: float
            alert_1.0%_recall:
              type: number
              format: float

    FeatureVector:
      type: object
      required:
        - Amount
        - log_amount
        - hour_sin
        - hour_cos
        - dow_sin
        - dow_cos
        - month_sin
        - month_cos
        - is_weekend
        - is_night
        - currency_mismatch
        - cross_border
        - is_round_amount
        - sender_txn_count_24h
        - sender_amount_sum_24h
        - sender_amount_mean_30d
        - sender_amount_std_30d
        - sender_amount_zscore
        - receiver_txn_count_24h
        - receiver_amount_sum_24h
        - seconds_since_sender_txn
        - sender_txn_count_lifetime
        - receiver_txn_count_lifetime
        - sender_out_degree
        - receiver_in_degree
        - pair_transaction_count
        - sender_counterparty_hhi
        - Payment_type
        - Payment_currency
        - Received_currency
        - Sender_bank_location
        - Receiver_bank_location
      properties:
        Amount:
          type: number
          format: float
          minimum: 0
          description: Transaction amount in base currency
        log_amount:
          type: number
          format: float
          description: Log(1 + Amount)
        hour_sin:
          type: number
          format: float
          minimum: -1
          maximum: 1
          description: Cyclical encoding of hour
        hour_cos:
          type: number
          format: float
          minimum: -1
          maximum: 1
        dow_sin:
          type: number
          format: float
          minimum: -1
          maximum: 1
          description: Day of week cyclical encoding
        dow_cos:
          type: number
          format: float
          minimum: -1
          maximum: 1
        month_sin:
          type: number
          format: float
          minimum: -1
          maximum: 1
          description: Month cyclical encoding
        month_cos:
          type: number
          format: float
          minimum: -1
          maximum: 1
        is_weekend:
          type: integer
          enum: [0, 1]
        is_night:
          type: integer
          enum: [0, 1]
        currency_mismatch:
          type: integer
          enum: [0, 1]
          description: 1 if payment_currency != received_currency
        cross_border:
          type: integer
          enum: [0, 1]
          description: 1 if sender_location != receiver_location
        is_round_amount:
          type: integer
          enum: [0, 1]
          description: 1 if amount is divisible by 1000
        sender_txn_count_24h:
          type: integer
          minimum: 0
        sender_amount_sum_24h:
          type: number
          format: float
          minimum: 0
        sender_amount_mean_30d:
          type: number
          format: float
          nullable: true
        sender_amount_std_30d:
          type: number
          format: float
          nullable: true
        sender_amount_zscore:
          type: number
          format: float
        receiver_txn_count_24h:
          type: integer
          minimum: 0
        receiver_amount_sum_24h:
          type: number
          format: float
          minimum: 0
        seconds_since_sender_txn:
          type: integer
          description: Time since sender's last transaction (-1 if none)
        sender_txn_count_lifetime:
          type: integer
          minimum: 0
        receiver_txn_count_lifetime:
          type: integer
          minimum: 0
        sender_out_degree:
          type: integer
          minimum: 0
          description: Unique counterparties sender has transacted with
        receiver_in_degree:
          type: integer
          minimum: 0
          description: Unique senders who have transacted with receiver
        pair_transaction_count:
          type: integer
          minimum: 0
          description: Number of prior transactions between sender-receiver pair
        sender_counterparty_hhi:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: HHI concentration measure for sender's counterparties
        Payment_type:
          type: string
          example: "Transfer"
        Payment_currency:
          type: string
          example: "USD"
        Received_currency:
          type: string
          example: "USD"
        Sender_bank_location:
          type: string
          example: "US"
        Receiver_bank_location:
          type: string
          example: "US"

    PredictionResponse:
      type: object
      required:
        - risk_probability
        - requires_review
      properties:
        risk_probability:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Calibrated risk probability (0 = low risk, 1 = high risk)
        requires_review:
          type: boolean
          description: Whether transaction exceeds alert threshold
        risk_percentile:
          type: number
          format: float
          minimum: 0
          maximum: 100
          nullable: true
          description: Percentile rank against validation set (0-100)

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing what went wrong
        detail:
          type: object
          nullable: true
          description: Additional error details (e.g., validation errors per field)